Massive performance optimizations to pattern generation.
Previous pattern generation was O(N+P*(M+N*log(N))) time complexity where N is the number of events in a chart, P is the number of patterns, and M is the number of steps in a pattern. And that N was a handful of O(N) scans, not just one. The algorithm was roughly:
Generate a new Event Tree: O(N)
For every pattern being generated: O(P)
	Create an ExpressedChart for the entire chart: O(N)
	Scan the ExpressedChart to determine footing and transitions: O(N)
	Create a PerformedChart: O(M)
	PerformedChart creation involved determining NPS and timing information: O(N)
	Create StepMania events from the PerformedChart: O(M)
	Set the time of the StepMania events by sorting them in with existing events: O(N*log(N))

This was obviously inefficient, but was relatively simple and correct in extreme edge cases for determining footing. For example there can be charts where you don't actually know the proper footing without scanning all the way back up to the start of the chart. For example consider a singles chart that starts with L R U D, then repeats U D for a long time going into a pattern. At the time of the pattern we don't know which foot is on U and which is on D without scanning back to the start of the chart. Scanning all the way to the start of the chart ensures we don't misinterpret this footing, but it is inefficient and these cases are incredibly rare. Similarly, scanning all the way to the end of the chart prevents similar issues with following footing.

Despite handling this edge case this old method introduced another arguably worse behavior. When using one ExpressedChart for the entire chart we would incorrectly try to connect regions of the chart that were separated by patterns. This caused some mistakes with determining the footing.

This old algorithm also waiting until all patterns were generated before committing results. So the user had to wait for all of them complete before seeing any new steps.

The new algorithm relies on small fixed-sized ranges, R (32), for scanning footing and involves only scanning a number of steps up to the start of the last pattern, S.

The new algorithm is roughly:
Get the timing events. Most charts have only a few. This is, in practice, a negligible amount of time.
Get the NPS from cached data: O(1)
For every pattern being generated:
	Scan from the previous pattern only up to this one for updating counts.
	Get a finite range of steps after the pattern: O(R+log(N))
	Create an ExpressedChart from this range: O(R)
	Check this ExpressedChart for the following footing: O(1)
	Get a finite range of steps before the pattern: O(R+log(N))
	Create an ExpressedChart from this range: O(R)
	Check this ExpressedChart for the previous footing and transitions: O(R)
	Create a PerformedChart: O(M)
	Create StepMania events from the PerformedChart: O(M)

This algorithm's time complexity is O(S+P*(R+log(N)+M)). This new algorithm will not handle the same extreme edge cases from the old algorithm, but it is orders of magnitude faster. It also improves footing determination by splitting up ExpressedCharts into one for the preceding footing and one for the following footing so that they are considered independently. It also commits results after each pattern so the user sees new steps come in incrementally.

In order to support the new algorithm the following changes were made:
- PerformedChart now stores the times of the events it generates. This lets us avoid additional sorts/scans to set the time.
- EditorActions now support an Update method which is called from the main update loop. This allows async pattern generation work to periodically commit results while it is still running.
- Added IsStep/IsLaneNote/IsConsumedByReceptors methods to EditorEvent to remove copy/paste code around checking EditorEvent types.