Songs now have an explicit TimingChart which is the chart used to set timing data on the Song when saving to sm or ssc files. This change is motivated primarily by a bug in Stepmania where even when having valid charts with complete timing data, Stepmania will still use the Song's timing data for some beat-driven animations like the receptor pulsing and cursor bouncing on the song wheel. When there is no Song timing data present it will use default values, causing these animations to play at 60bpm. To address this we need to set timing data on the Song even when using split timing in ssc files.
This change also simplifies sm file saving. Now that there is an explicit TimingChart, we can use that all the time when saving sm files and warn on timing incompatibilities rather than treating this as an error and failing to save.

This commit also fixes a bug where saving wouldn't clear the unsaved changes flag when it fails.